+++ cobalt/editor/editor.h
@@ -16,7 +16,7 @@ namespace cobalt {
 
             void fixedTimeStep() override;
             void variableTimeStep(const float delta) override;
-            void onResize(const float width, const float height) override;
+            void onResize(const uint width, const uint height) override;
             void bindInput();
             void createScene();
